[docker] Add docker with demo file
Add a dockerfile which installs the prereqs.
When built and ran, User will automatically enter a tmux session with FAN predicting the Astronaut sklearn image
When built and ran, User will automatically enter a tmux session with FAN predicting the Astronaut sklearn image, this behaviour can be changed in future commits.

Detector used for now is simply the HaarCascade from opencv.

Next steps:
- Either build a flask server with UploadImage button, detects the face and draws the landmarks
- Or a jupyter notebook with a similar functionality
- Script the model, currently I cannot do it because _modules is an issue in HourGlass
- I haven't yet tested on GPUs, I believe I need to detect the gpu in bash and pass them using --gpus all